+++ src/main/java/com/invoice/repository/BillingRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.invoice.entity.BillingHeader;
 
 @Repository
-public interface BillingRepository extends JpaRepository<BillingHeader, Long> {
+public interface BillingRepository extends JpaRepository<BillingHeader, Integer> {
 
 }
